requirements.txt
crewai==0.80.0
streamlit==1.41.1
langchain-openai==0.2.14
openai==1.58.1
python-dotenv==1.0.1
pydantic==2.10.4
